+++ changelog.md
@@ -1,3 +1,7 @@
-# 1.0.0
-- Initial release.
-- If you're seeing this on the geode index, just know that I wrote this text when I first started making mods, and somehow I never touched the changelog file ever again
+# v0.1.0
+- First test release
+# v0.2.0
+- Added a logo to the mod
+- Added a name
+- Added a test jumpscare button to the settings menu
+- Reorganized a bunch of code